+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added visualization type selector for LED lights. The new overlay marker based
   visualization can speed up playback and render speed when using light effects.
 
+- Added marker size setting for the MARKERS type pyro and LED visualization.
+
 - Added audio export option to the .skyc exporter, requiring a single .mp3 music
   file to be added as an audio strip in the VSE.
 
@@ -51,6 +53,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed an exception that happened when switching between files containing video based
   light effects.
+
+- Pyro overlay markers are not drawn any more when visualization type is INFO.
 
 ### Removed